refactor(cli): migrate cmems tooling onto discovery
Move cmems's refresher/writer/prober/deep_prober to earthlens.cmems.cli, wiring
the first deep_prober role through discovery (curate's _DEEP_PROBERS now projects
dispatch_table). Relocate cmems's CLI tests to earthlens-ocean. Shrink the guard
allow-list (cmems).
